refactor: locate reka chat template handler in chat.cpp
